enhance interface
