--- src/constants/report_table_data.ts
@@ -1,6 +1,0 @@
-import {IReportTable} from '../interfaces/report_table';
-
-export const reportsDateSpan = {
-  start: 1688190893,
-  end: 1690646400,
-};
